Remove call to function url() for each page
Profiling with xhprof shows that the generating the url with the
function url() is not fast, especially when making use of modules like
pathauto. When a book contains more than 100 pages the function url()
is also called 100 times, and if each call takes 0,01 seconds this will
take a full second. Larger books have even more problems.
By replacing the call to function url() with a simple string concatenation
the loading times of books can be cut in half.